feature: NAPT実装できたっぽい. 名前解決はできないっぽい

+++ base.h
@@ -42,7 +42,7 @@ struct pseudo_ip {
   struct in_addr ip_dst;
   uint8_t dummy;
   uint8_t ip_p;
-  uint8_t ip_len;
+  uint16_t ip_len;
 };
 
 #endif /* BASE_H */
